ui: take responsibility from output
