+++ src/libtrx/game/output/draw.c
@@ -2,7 +2,6 @@
 
 #include "config.h"
 #include "game/creature/const.h"
-#include "game/objects.h"
 #include "game/output.h"
 #include "game/output/sources/lightnings.h"
 #include "game/output/sources/misc.h"
@@ -16,99 +15,6 @@
 #include "game/scaler.h"
 #include "game/shell.h"
 #include "version.h"
-
-#define M_OUTLINE_THICKNESS 0.75f
-
-typedef enum {
-    C_BACKGROUND_E,
-    C_BACKGROUND_C,
-    C_BACKGROUND_HEAVY_E,
-    C_BACKGROUND_HEAVY_C,
-    C_HEADING_E,
-    C_HEADING_C,
-    C_REQUESTED_E,
-    C_REQUESTED_C,
-    C_REQUESTED_OUTLINE_CH,
-    C_REQUESTED_OUTLINE_CV,
-    C_REQUESTED_OUTLINE_E,
-    C_BACKGROUND_OUTLINE_TL,
-    C_BACKGROUND_OUTLINE_TR,
-    C_BACKGROUND_OUTLINE_BL,
-    C_BACKGROUND_OUTLINE_BR,
-    C_HEADING_OUTLINE,
-    C_GENERIC_OUTLINE_LIGHT,
-    C_GENERIC_OUTLINE_DARK,
-    C_NUMBER_OF,
-} M_COLOR;
-
-typedef union {
-    M_COLOR colors[2];
-    struct {
-        M_COLOR edge;
-        M_COLOR center;
-    };
-} M_GRADIENT_FILL;
-
-static M_GRADIENT_FILL m_GradientFills[] = {
-    // clang-format off
-    [TS_BACKGROUND]       = { .edge = C_BACKGROUND_E, .center = C_BACKGROUND_C },
-    [TS_BACKGROUND_HEAVY] = { .edge = C_BACKGROUND_HEAVY_E, .center = C_BACKGROUND_HEAVY_C },
-    [TS_HEADING]          = { .edge = C_HEADING_E, .center = C_HEADING_C },
-    [TS_REQUESTED]        = { .edge = C_REQUESTED_E, .center = C_REQUESTED_C },
-    [TS_LINE]             = { .edge = C_BACKGROUND_E, .center = C_BACKGROUND_C },
-    // clang-format on
-};
-
-static RGBA_8888 m_MenuColorMap[TR_VERSION_COUNT][C_NUMBER_OF] = {
-    // clang-format off
-    [0] = {
-        [C_BACKGROUND_C]           = { 0x00, 0x00, 0x40, 0x80 },
-        [C_BACKGROUND_E]           = { 0x00, 0x00, 0x00, 0x80 },
-        [C_BACKGROUND_HEAVY_C]     = { 0x00, 0x00, 0x00, 0xE0 },
-        [C_BACKGROUND_HEAVY_E]     = { 0x00, 0x00, 0x00, 0xE0 },
-        [C_HEADING_E]              = { 0x00, 0x00, 0x00, 0x80 },
-        [C_HEADING_C]              = { 0x80, 0x38, 0x10, 0x80 },
-        [C_REQUESTED_E]            = { 0x00, 0x00, 0x00, 0x80 },
-        [C_REQUESTED_C]            = { 0x80, 0x38, 0xDC, 0x80 },
-        //[C_REQUESTED_C]            = { 0x40, 0x1C, 0x78, 0x80 },
-        [C_REQUESTED_OUTLINE_CH]   = { 0xC8, 0xC8, 0xC8, 0xFF },
-        [C_REQUESTED_OUTLINE_CV]   = { 0xC8, 0xC8, 0xC8, 0xFF },
-        [C_REQUESTED_OUTLINE_E]    = { 0x28, 0x28, 0x28, 0xFF },
-        [C_BACKGROUND_OUTLINE_TL]  = { 0x60, 0x60, 0x60, 0xFF },
-        [C_BACKGROUND_OUTLINE_TR]  = { 0x20, 0x20, 0x20, 0xFF },
-        [C_BACKGROUND_OUTLINE_BL]  = { 0x40, 0x40, 0x40, 0xFF },
-        [C_BACKGROUND_OUTLINE_BR]  = { 0x00, 0x00, 0x00, 0xFF },
-        [C_HEADING_OUTLINE]        = { 0x00, 0x00, 0x00, 0xFF },
-        [C_GENERIC_OUTLINE_LIGHT]  = { 0xE8, 0xC0, 0x70, 0xFF },
-        [C_GENERIC_OUTLINE_DARK]   = { 0x8C, 0x70, 0x38, 0xFF },
-    },
-    [1] = {
-        [C_BACKGROUND_E]           = { 0x00, 0x20, 0x00, 0x80 },
-        [C_BACKGROUND_C]           = { 0x00, 0x60, 0x00, 0x80 },
-        [C_BACKGROUND_HEAVY_E]     = { 0x00, 0x00, 0x00, 0xE0 },
-        [C_BACKGROUND_HEAVY_C]     = { 0x00, 0x20, 0x00, 0xE0 },
-        [C_HEADING_E]              = { 0x00, 0x00, 0x00, 0x80 },
-        [C_HEADING_C]              = { 0x10, 0x80, 0x38, 0x80 },
-        [C_REQUESTED_E]            = { 0x00, 0x00, 0x00, 0x80 },
-        [C_REQUESTED_C]            = { 0x38, 0xF0, 0x80, 0x80 },
-        [C_REQUESTED_OUTLINE_CH]   = { 0xFF, 0xFF, 0xFF, 0xFF },
-        [C_REQUESTED_OUTLINE_CV]   = { 0x38, 0xF0, 0x80, 0xFF },
-        [C_REQUESTED_OUTLINE_E]    = { 0x00, 0x00, 0x00, 0xFF },
-        [C_BACKGROUND_OUTLINE_TL]  = { 0x60, 0x60, 0x60, 0xFF },
-        [C_BACKGROUND_OUTLINE_TR]  = { 0x20, 0x20, 0x20, 0xFF },
-        [C_BACKGROUND_OUTLINE_BL]  = { 0x40, 0x40, 0x40, 0xFF },
-        [C_BACKGROUND_OUTLINE_BR]  = { 0x00, 0x00, 0x00, 0xFF },
-        [C_HEADING_OUTLINE]        = { 0x00, 0x00, 0x00, 0xFF },
-        [C_GENERIC_OUTLINE_LIGHT]  = { 0xE8, 0xC0, 0x70, 0xFF },
-        [C_GENERIC_OUTLINE_DARK]   = { 0x8C, 0x70, 0x38, 0xFF },
-    },
-    // clang-format on
-};
-
-static RGBA_8888 M_GetMenuColor(const M_COLOR color)
-{
-    return m_MenuColorMap[g_TRVersion - 1][color];
-}
 
 static void M_DrawScreenQuad(
     const int32_t x0, const int32_t y0, const int32_t x1, const int32_t y1,
@@ -235,130 +141,6 @@ void Output_DrawScreenSprite(
     });
 }
 
-void Output_DrawTextBackground(
-    const UI_STYLE ui_style, const int32_t sx, const int32_t sy,
-    const int32_t z, int32_t w, int32_t h, const TEXT_STYLE text_style)
-{
-    // Make sure height and width divisible by 2.
-    w = 2 * ((w + 1) / 2);
-    h = 2 * ((h + 1) / 2);
-
-    if (ui_style == UI_STYLE_PC) {
-        const RGBA_8888 cb = { 0, 0, 0,
-                               text_style == TS_BACKGROUND_HEAVY ? 224 : 128 };
-        M_DrawScreenQuad(sx, sy, sx + w, sy + h, z, cb, cb, cb, cb);
-        return;
-    }
-
-    const int32_t x0 = sx;
-    const int32_t y0 = sy;
-    const int32_t x1 = sx + w;
-    const int32_t y1 = sy + h;
-    const int32_t xm = sx + w / 2;
-    const int32_t ym = sy + h / 2;
-
-    const M_GRADIENT_FILL *const fill = &m_GradientFills[text_style];
-#define L_DRAW(x0, y0, x1, y1, tl, tr, bl, br)                                 \
-    M_DrawScreenQuad(                                                          \
-        x0, y0, x1, y1, z, M_GetMenuColor(fill->colors[tl]),                   \
-        M_GetMenuColor(fill->colors[tr]), M_GetMenuColor(fill->colors[bl]),    \
-        M_GetMenuColor(fill->colors[br]));
-    L_DRAW(xm, y0, x0, ym, 0, 0, 1, 0);
-    L_DRAW(x1, y0, xm, ym, 0, 0, 0, 1);
-    L_DRAW(xm, ym, x0, y1, 1, 0, 0, 0);
-    L_DRAW(x1, ym, xm, y1, 0, 1, 0, 0);
-#undef L_DRAW
-}
-
-void Output_DrawTextOutline(
-    const UI_STYLE ui_style, const int32_t sx, const int32_t sy,
-    const int32_t z, int32_t w, int32_t h, const TEXT_STYLE text_style)
-{
-    if (g_TRVersion == 2 && ui_style == UI_STYLE_PC) {
-        const int32_t mesh_idx = Object_Get(O_TEXT_BOX)->mesh_idx;
-
-        const int32_t offset = text_style == TS_LINE ? 0 : 4;
-        const int32_t x0 = sx + offset;
-        const int32_t y0 = sy + offset;
-        const int32_t x1 = x0 + w - offset * 2;
-        const int32_t y1 = y0 + h - offset * 2;
-        const int32_t scale_h = PHD_ONE;
-        const int32_t scale_v = PHD_ONE;
-        w = (w - offset * 2) * PHD_ONE / 8;
-        h = (h - offset * 2) * PHD_ONE / 8;
-
-        if (text_style == TS_LINE) {
-            Output_DrawScreenSprite(
-                x0, y0, z, w, scale_v, mesh_idx + 4, SHADE_NEUTRAL);
-        } else {
-            // Corners
-            Output_DrawScreenSprite(
-                x0, y0, z, scale_h, scale_v, mesh_idx + 0, SHADE_NEUTRAL);
-            Output_DrawScreenSprite(
-                x1, y0, z, scale_h, scale_v, mesh_idx + 1, SHADE_NEUTRAL);
-            Output_DrawScreenSprite(
-                x1, y1, z, scale_h, scale_v, mesh_idx + 2, SHADE_NEUTRAL);
-            Output_DrawScreenSprite(
-                x0, y1, z, scale_h, scale_v, mesh_idx + 3, SHADE_NEUTRAL);
-
-            // Lines
-            Output_DrawScreenSprite(
-                x0, y0, z, w, scale_v, mesh_idx + 4, SHADE_NEUTRAL);
-            Output_DrawScreenSprite(
-                x1, y0, z, scale_h, h, mesh_idx + 5, SHADE_NEUTRAL);
-            Output_DrawScreenSprite(
-                x0, y1, z, w, scale_v, mesh_idx + 6, SHADE_NEUTRAL);
-            Output_DrawScreenSprite(
-                x0, y0, z, scale_h, h, mesh_idx + 7, SHADE_NEUTRAL);
-        }
-        return;
-    }
-
-    if (g_TRVersion == 1 && ui_style == UI_STYLE_PC) {
-        const RGBA_8888 cd = M_GetMenuColor(C_GENERIC_OUTLINE_DARK);
-        const RGBA_8888 cl = M_GetMenuColor(C_GENERIC_OUTLINE_LIGHT);
-        const float thickness = M_OUTLINE_THICKNESS;
-        if (text_style == TS_LINE) {
-            const float e = Scaler_Calc(thickness, SCALER_TARGET_TEXT);
-            const float cy = sy + h / 2;
-            M_DrawScreenQuad(sx, cy - e, sx + w, cy, 0, cl, cl, cl, cl);
-            M_DrawScreenQuad(sx, cy, sx + w, cy + e, 0, cd, cd, cd, cd);
-        } else {
-            Output_DrawScreenFrame(sx, sy, w, h, cd, cl, thickness);
-        }
-        return;
-    }
-
-    if (text_style == TS_HEADING) {
-        Output_DrawScreenGradientBox(
-            sx, sy, w, h, M_GetMenuColor(C_HEADING_OUTLINE),
-            M_GetMenuColor(C_HEADING_OUTLINE),
-            M_GetMenuColor(C_HEADING_OUTLINE),
-            M_GetMenuColor(C_HEADING_OUTLINE), M_OUTLINE_THICKNESS);
-    } else if (
-        text_style == TS_BACKGROUND || text_style == TS_BACKGROUND_HEAVY) {
-        Output_DrawScreenGradientBox(
-            sx, sy, w, h, M_GetMenuColor(C_BACKGROUND_OUTLINE_TL),
-            M_GetMenuColor(C_BACKGROUND_OUTLINE_TR),
-            M_GetMenuColor(C_BACKGROUND_OUTLINE_BL),
-            M_GetMenuColor(C_BACKGROUND_OUTLINE_BR), M_OUTLINE_THICKNESS);
-    } else if (text_style == TS_LINE) {
-        M_DrawScreenQuad(
-            sx, sy, sx + w, sy + h, 0, M_GetMenuColor(C_BACKGROUND_OUTLINE_TL),
-            M_GetMenuColor(C_BACKGROUND_OUTLINE_TR),
-            M_GetMenuColor(C_BACKGROUND_OUTLINE_BL),
-            M_GetMenuColor(C_BACKGROUND_OUTLINE_BR));
-    } else if (text_style == TS_REQUESTED) {
-        // Make sure height and width divisible by 2.
-        w = 2 * ((w + 1) / 2);
-        h = 2 * ((h + 1) / 2);
-        Output_DrawScreenCentreGradientBox(
-            sx, sy, w, h, M_GetMenuColor(C_REQUESTED_OUTLINE_E),
-            M_GetMenuColor(C_REQUESTED_OUTLINE_CH),
-            M_GetMenuColor(C_REQUESTED_OUTLINE_CV), M_OUTLINE_THICKNESS);
-    }
-}
-
 void Output_DrawScreenFlatQuad(
     const int32_t sx, const int32_t sy, const int32_t z, const int32_t w,
     const int32_t h, const RGBA_8888 color)
@@ -372,51 +154,6 @@ void Output_DrawScreenGradientQuad(
     const RGBA_8888 br)
 {
     M_DrawScreenQuad(sx, sy, sx + w, sy + h, z, tl, tr, bl, br);
-}
-
-void Output_DrawScreenGradientBox(
-    const int32_t sx, const int32_t sy, const int32_t w, const int32_t h,
-    const RGBA_8888 tl, const RGBA_8888 tr, const RGBA_8888 bl,
-    const RGBA_8888 br, const float thickness)
-{
-    const float e = Scaler_Calc(thickness, SCALER_TARGET_TEXT);
-    const float x0 = sx;
-    const float y0 = sy;
-    const float x1 = sx + w;
-    const float y1 = sy + h;
-    M_DrawScreenQuad(x0 - e, y0 - e, x1 + e, y0 + e, 0, tl, tr, tl, tr);
-    M_DrawScreenQuad(x0 - e, y1 - e, x1 + e, y1 + e, 0, bl, br, bl, br);
-    M_DrawScreenQuad(x0 - e, y0 - e, x0 + e, y1 + e, 0, tl, tl, bl, bl);
-    M_DrawScreenQuad(x1 - e, y0 - e, x1 + e, y1 + e, 0, tr, tr, br, br);
-}
-
-void Output_DrawScreenCentreGradientBox(
-    const int32_t sx, const int32_t sy, const int32_t w, const int32_t h,
-    const RGBA_8888 edge, const RGBA_8888 center_h, const RGBA_8888 center_v,
-    const float thickness)
-{
-    const float e = Scaler_Calc(thickness, SCALER_TARGET_TEXT);
-    const float x0 = sx;
-    const float y0 = sy;
-    const float x1 = sx + w;
-    const float y1 = sy + h;
-    const float xm = (x0 + x1) / 2.0f;
-    const float ym = (y0 + y1) / 2.0f;
-    const RGBA_8888 ch = center_h;
-    const RGBA_8888 cv = center_v;
-    const RGBA_8888 ce = edge;
-    const RGBA_8888 cc = (RGBA_8888) { 255, 0, 0, 255 };
-
-    // clang-format off
-    M_DrawScreenQuad(x0 - e, y0 - e, xm,     y0 + e, 0, ce, ch, ce, ch);
-    M_DrawScreenQuad(xm,     y0 - e, x1 + e, y0 + e, 0, ch, ce, ch, ce);
-    M_DrawScreenQuad(x0 - e, y1 - e, xm,     y1 + e, 0, ce, ch, ce, ch);
-    M_DrawScreenQuad(xm,     y1 - e, x1 + e, y1 + e, 0, ch, ce, ch, ce);
-    M_DrawScreenQuad(x0 - e, y0,     x0 + e, ym,     0, ce, ce, cv, cv);
-    M_DrawScreenQuad(x0 - e, ym,     x0 + e, y1,     0, cv, cv, ce, ce);
-    M_DrawScreenQuad(x1 - e, y0,     x1 + e, ym,     0, ce, ce, cv, cv);
-    M_DrawScreenQuad(x1 - e, ym,     x1 + e, y1,     0, cv, cv, ce, ce);
-    // clang-format on
 }
 
 void Output_DrawScreenFrame(
