+++ src/libtrx/include/libtrx/version.h
@@ -2,5 +2,7 @@
 
 #include <stdint.h>
 
+#define TR_VERSION_COUNT 2
+
 extern const char *g_TRXVersion;
 extern int32_t g_TRVersion;
